Add prototype player move force slider bar

+++ src/components/player.hpp
@@ -5,7 +5,7 @@ class Player
 public:
 	bool can_switch_charge = true;
 	bool can_go_neutral = true;
-	float move_force = 0;
+	int move_force = 0;
 };
 
 class DeadPlayer
